+++ docs/CHANGELOG.md
@@ -6,6 +6,8 @@
 When u put these tools so close together u start to see unpreceedented ways to combine them
 
 - better way for polyglot dev: when u test things in nodejs and py too. Need to start server easily and fast and run things
+- finish ner, use with KB
+- make io_start return then for chaining
 - start first neural net to simple-classify sentence into intents
 - better menu format inspired by lita
 - skflow gridsearch hyperparam with spark https://databricks.com/blog/2016/02/08/auto-scaling-scikit-learn-with-spark.html
@@ -88,3 +90,4 @@ When u put these tools so close together u start to see unpreceedented ways to c
 - promisify all ai modules for chaining
 - add google kg search with interface
 - restore py `Textblob`
+- make io_start return promise for chaining